Fix CSRF/Stored XSS vulnerability
Add X-CSRF token (wpnonce) to administration edit form.  This fixes an issue that would allow an attacker to create a CSRF attack and overwrite the "hljs_additional_css" options with whatever injected HTML they desire, this HTML could escape out of the <style> tag with </style> and begin injecting <scripts> into the blog.  Since the CSS exists on all pages this injection would be persistent throughout the blog site.